+++ src/OpticalComponents/Components.jl
@@ -5,7 +5,7 @@ include("Lenses.jl")
 include("Prisms.jl")
 include("SphericalLenses.jl")
 include("DoubletLenses.jl")
-include("Detectors/Detectors.jl")
+include("Detectors/Detector.jl")
 include("Beamsplitters/Beamsplitters.jl")
 include("NonInteractable.jl")
 include("Intersectable.jl")
