+++ src/OpticalComponents/Detectors/FieldCalculation.jl
